Move module types to types dir
