+++ packages/chookscord/src/types/index.ts
@@ -2,3 +2,4 @@ import './context';
 import './env';
 
 export * from './config';
+export * from './modules';
